ny version af moduler samt tests

+++ integrationtest/test/src/test/java/dk/kvalitetsit/kitcaddy/test/AllInOneExpiryIntegrationTest.java
@@ -1,16 +1,23 @@
 package dk.kvalitetsit.kitcaddy.test;
 
+import java.io.IOException;
 import java.util.UUID;
+import java.util.function.Predicate;
 
 import org.json.JSONException;
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
+import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
+import org.testcontainers.containers.wait.strategy.Wait;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -21,7 +28,7 @@ import com.mongodb.client.result.DeleteResult;
 import dk.kvalitetsit.kitcaddy.AbstractAllInOneIT;
 import dk.kvalitetsit.kitcaddy.TestConstants;
 
-public class AllInOneExpiryIT extends AbstractAllInOneIT {
+public class AllInOneExpiryIntegrationTest extends AbstractAllInOneIT {
 
 	private static final String TEST_URL = "http://"+SAML_SP_URL+"/service/test";
 
@@ -66,6 +73,64 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 	}
 
 	@Test
+	@Ignore
+	public void testMongoRestartIsHandledTransparently() throws JSONException, InterruptedException  {
+		// Denne test virker, når jeg debugger den
+		
+		// Given
+		Expiry expiry = new Expiry() {
+			@Override
+			public void doExpiry(JsonNode resultBeforeExpiry) {
+				// This will also remove all data in mongo
+				mongoContainer.stop();
+				mongoContainer.start();
+			}
+		};
+
+		// When
+		Response responseAfterExpiry = resultAfterExpiry(expiry, false); // Do not remove keycloak sessions as we want to "autologin"		
+		Thread.sleep(8000); // Let the loaded pages calm down
+		
+		// Then
+		JsonNode responseJsonAfterExpiryParsed = parseJsonReturned(responseAfterExpiry.getWebDriver().getPageSource());
+		Assert.assertNotNull("Expected a json response", responseJsonAfterExpiryParsed);
+
+
+		String wspAuthorizationHeaderBeforeExpiry = ((TextNode) responseAfterExpiry.getResponseBeforeExpiry().get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
+		String wspAuthorizationHeaderAfterExpiry = ((TextNode) responseJsonAfterExpiryParsed.get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
+		Assert.assertNotEquals("Expected a new session on the WSP - checking that the authorization header has changed", wspAuthorizationHeaderBeforeExpiry, wspAuthorizationHeaderAfterExpiry);
+	}
+
+	@Test
+	public void testMongoConnectionExpiryIsHandledTransparently() throws JSONException  {
+
+		// Given
+		Expiry expiry = new Expiry() {
+			@Override
+			public void doExpiry(JsonNode resultBeforeExpiry) {
+				try {
+					mongoContainer.execInContainer("mongo < /scripts/killallconnections.js");
+				} catch (UnsupportedOperationException | IOException | InterruptedException e) {
+					throw new RuntimeException(e);
+				}
+			}
+		};
+
+		// When
+		Response responseAfterExpiry = resultAfterExpiry(expiry);		
+
+		// Then
+		JsonNode responseJsonAfterExpiryParsed = parseJsonReturned(responseAfterExpiry.getWebDriver().getPageSource());
+		Assert.assertNotNull("Expected a json response", responseJsonAfterExpiryParsed);
+
+
+		String wspAuthorizationHeaderBeforeExpiry = ((TextNode) responseAfterExpiry.getResponseBeforeExpiry().get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
+		String wspAuthorizationHeaderAfterExpiry = ((TextNode) responseJsonAfterExpiryParsed.get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
+		Assert.assertEquals("Expected the same session on the WSP - checking that the authorization header are the same", wspAuthorizationHeaderBeforeExpiry, wspAuthorizationHeaderAfterExpiry);
+	}
+
+
+	@Test
 	public void testWscSessionExpiryIsHandledTransparently() throws JSONException {
 
 		// Given
@@ -87,8 +152,8 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 		// Then
 		JsonNode responseJsonAfterExpiryParsed = parseJsonReturned(responseAfterExpiry.getWebDriver().getPageSource());
 		Assert.assertNotNull("Expected a json response", responseJsonAfterExpiryParsed);
-		
-		
+
+
 		String wspAuthorizationHeaderBeforeExpiry = ((TextNode) responseAfterExpiry.getResponseBeforeExpiry().get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
 		String wspAuthorizationHeaderAfterExpiry = ((TextNode) responseJsonAfterExpiryParsed.get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
 		Assert.assertNotEquals("Expected a new session on the WSP - checking that the authorization header has changed", wspAuthorizationHeaderBeforeExpiry, wspAuthorizationHeaderAfterExpiry);
@@ -117,7 +182,7 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 		String responseAfterExpiryBody = responseAfterExpiry.getWebDriver().getPageSource();
 		JsonNode responseJsonAfterExpiryParsed = parseJsonReturned(responseAfterExpiryBody);
 		Assert.assertNotNull("Expected a json response", responseJsonAfterExpiryParsed);
-		
+
 		String wspAuthorizationHeaderBeforeExpiry = ((TextNode) responseAfterExpiry.getResponseBeforeExpiry().get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
 		String wspAuthorizationHeaderAfterExpiry = ((TextNode) responseJsonAfterExpiryParsed.get(TestConstants.ECHO_SERVICE_HTTP_HEADER_KEY).get(TestConstants.WSP_AUTHORIZATION_HEADERNAME)).textValue();
 		Assert.assertNotEquals("Expected a new session on the WSP - checking that the authorization header has changed", wspAuthorizationHeaderBeforeExpiry, wspAuthorizationHeaderAfterExpiry);
@@ -125,6 +190,10 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 
 
 	private Response resultAfterExpiry(Expiry expiry) throws JSONException  {
+		return resultAfterExpiry(expiry, true);
+	}
+
+	private Response resultAfterExpiry(Expiry expiry, boolean removeKeyCloakCookiesBeforeExpiry) throws JSONException  {
 		// Perform login
 		String username = UUID.randomUUID().toString();
 		String password = "secret1234";
@@ -135,10 +204,12 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 		// Access protected ressource
 		JsonNode responseParsed = parseJsonReturned(resultBeforeExpiry);
 
-		// Make sure that we don't log automatically into keycloak on session expiry
-		webdriver.get(TestConstants.KEYCLOAK_ACCOUNT_URL);
-		webdriver.getPageSource();
-		webdriver.manage().deleteAllCookies();
+		if (removeKeyCloakCookiesBeforeExpiry) {
+			// Make sure that we don't log automatically into keycloak on session expiry
+			webdriver.get(TestConstants.KEYCLOAK_ACCOUNT_URL);
+			webdriver.getPageSource();
+			webdriver.manage().deleteAllCookies();
+		}
 
 		// Expire session
 		expiry.doExpiry(responseParsed);
@@ -152,6 +223,7 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 	private JsonNode parseJsonReturned(String result) {
 
 		if (!result.contains("{") || !result.contains("}")) {
+			System.out.println("Not parsed: "+result);
 			return null;
 		}
 
@@ -160,6 +232,7 @@ public class AllInOneExpiryIT extends AbstractAllInOneIT {
 			JsonNode responseParsed = new ObjectMapper().readValue(jsonReturned, JsonNode.class);
 			return responseParsed;
 		} catch (JsonProcessingException e) {
+			System.out.println("Not parsed: "+result);
 			return null;
 		}
 	}
